# perf(scanner): update type info in worker threads