--- a/crates/biome_service/src/workspace/scanner.rs
+++ b/crates/biome_service/src/workspace/scanner.rs
@@ -8,12 +8,10 @@
 //! In other words, the scanner is both the scanning logic in this module as
 //! well as the watcher to allow continuous scanning.
 
-use super::ServiceDataNotification;
 use super::server::WorkspaceServer;
 use crate::diagnostics::Panic;
 use crate::projects::ProjectKey;
 use crate::workspace::{DocumentFileSource, FileContent, OpenFileParams};
-use crate::workspace_watcher::WatcherSignalKind;
 use crate::{Workspace, WorkspaceError};
 use biome_diagnostics::serde::Diagnostic;
 use biome_diagnostics::{Diagnostic as _, Error, Severity};
@@ -124,10 +122,6 @@ fn scan_folder(folder: &Utf8Path, ctx: ScanContext) -> Duration {
         }
     }));
 
-    let mut paths = configs;
-    paths.append(&mut manifests);
-    ctx.workspace
-        .update_project_layout_for_paths(WatcherSignalKind::AddedOrChanged, &paths);
     let result = ctx
         .workspace
         .update_project_ignore_files(ctx.project_key, &ignore_paths);
@@ -141,11 +135,6 @@ fn scan_folder(folder: &Utf8Path, ctx: ScanContext) -> Duration {
             scope.handle(ctx_ref, path.to_path_buf());
         }
     }));
-
-    let _ = ctx
-        .workspace
-        .notification_tx
-        .send(ServiceDataNotification::Updated);
 
     start.elapsed()
 }
@@ -290,9 +279,6 @@ impl TraversalContext for ScanContext<'_> {
 
     fn handle_path(&self, path: BiomePath) {
         open_file(self, &path);
-
-        self.workspace
-            .update_module_graph(WatcherSignalKind::AddedOrChanged, &[path]);
     }
 
     fn store_path(&self, path: BiomePath) {