## Summary

The scanner opens files in the worker threads powered by rayon, but collecting type info is run in the main thread after the traversal. It contains type inference, which is CPU-heavy operation. Plus, it can be moved to threads safely.

You can see only the main thread is busy while scanning in the following screenshot:

<img width="1669" alt="image" src="https://github.com/user-attachments/assets/c1d6c8ba-1dc0-40d3-9797-664013d698f1" />

During investigation, I noticed that `ctx.workspace.update_module_graph()` is also called in the traversal. It tried to resolve the imported modules to update the graph, but it always fail due to a small bug in the resolver cache. Finally I removed duplicated calls of `ctx.workspace.update_module_graph()`, so now the graph is updated in the first call during the traversal. After this, all worker threads are used to collecting module graph and infering types:

<img width="1664" alt="image" src="https://github.com/user-attachments/assets/c22fc75d-cdfa-4fc6-bb30-cf602d9cde5d" />


## Test Plan

Existing tests should pass.

Locally tested to see the execution time.

Before (from #5839):

```
❯ time ~/Projects/github.com/biomejs/biome/target/release-with-debug/biome check
Checked 5243 files in 6s. No fixes applied.

________________________________________________________
Executed in    7.36 secs    fish           external
   usr time   31.41 secs    0.24 millis   31.41 secs
   sys time   17.50 secs    2.30 millis   17.50 secs
```

After

```
❯ time ~/Projects/github.com/biomejs/biome/target/release-with-debug/biome check
Checked 5243 files in 5s. No fixes applied.

________________________________________________________
Executed in    6.33 secs    fish           external
   usr time   32.88 secs  105.00 micros   32.88 secs
   sys time   16.60 secs  783.00 micros   16.60 secs
```